# rowcontainer,rowexec: switch joinReader to use DiskBackedNumberedRowC…