…ontainer

Additionally,
- added a randomized correctness test that compares the results of the indexed
  and numbered containers.
- added benchmark cases to the joinReader benchmark that limit memory. None of the
  workloads have repeated reads of the same right row and all access the right
  rows in monotonically increasing order so the difference between the two
  containers is due to the numbered container avoiding the overhead of populating
  the cache.
- reduced the number of slice allocations in newNumberedDiskRowIterator.

Fixes #48118

Release note: None