@sumeerbhola 2020-05-28
benchstat output

```
name                                                                                                old time/op    new time/op     delta
JoinReader/reqOrdering=true/matchratio=onetoone/lookuprows=1/mem=100KB/parallel=true-16               45.1µs ± 1%     43.5µs ± 2%    -3.56%  (p=0.008 n=5+5)
JoinReader/reqOrdering=true/matchratio=onetoone/lookuprows=1/mem=100KB-16                             46.2µs ± 2%     44.3µs ± 2%    -4.02%  (p=0.016 n=5+5)
JoinReader/reqOrdering=true/matchratio=onetoone/lookuprows=1/mem=unlimited/parallel=true-16           45.9µs ± 0%     44.3µs ± 2%    -3.51%  (p=0.016 n=4+5)
JoinReader/reqOrdering=true/matchratio=onetoone/lookuprows=1/mem=unlimited-16                         45.9µs ± 1%     44.5µs ± 2%    -3.02%  (p=0.008 n=5+5)
JoinReader/reqOrdering=true/matchratio=onetoone/lookuprows=16/mem=100KB-16                             150µs ± 2%      149µs ± 1%      ~     (p=0.222 n=5+5)
JoinReader/reqOrdering=true/matchratio=onetoone/lookuprows=16/mem=unlimited-16                         150µs ± 2%      149µs ± 2%      ~     (p=0.690 n=5+5)
JoinReader/reqOrdering=true/matchratio=onetoone/lookuprows=256/mem=100KB-16                           1.80ms ± 1%     1.85ms ± 5%      ~     (p=0.095 n=5+5)
JoinReader/reqOrdering=true/matchratio=onetoone/lookuprows=256/mem=unlimited-16                       1.81ms ± 2%     1.79ms ± 2%      ~     (p=0.421 n=5+5)
JoinReader/reqOrdering=true/matchratio=onetoone/lookuprows=1024/mem=100KB-16                          7.03ms ± 2%     7.03ms ± 3%      ~     (p=1.000 n=5+5)
JoinReader/reqOrdering=true/matchratio=onetoone/lookuprows=1024/mem=unlimited-16                      7.21ms ± 3%     7.09ms ± 1%      ~     (p=0.222 n=5+5)
JoinReader/reqOrdering=true/matchratio=onetoone/lookuprows=4096/mem=100KB-16                          28.3ms ± 2%     28.0ms ± 1%      ~     (p=0.222 n=5+5)
JoinReader/reqOrdering=true/matchratio=onetoone/lookuprows=4096/mem=unlimited-16                      28.2ms ± 1%     28.1ms ± 3%      ~     (p=1.000 n=5+5)
JoinReader/reqOrdering=true/matchratio=onetoone/lookuprows=8192/mem=100KB-16                          56.9ms ± 2%     56.1ms ± 2%      ~     (p=0.151 n=5+5)
JoinReader/reqOrdering=true/matchratio=onetoone/lookuprows=8192/mem=unlimited-16                      56.8ms ± 3%     56.5ms ± 2%      ~     (p=1.000 n=5+5)
JoinReader/reqOrdering=true/matchratio=onetoone/lookuprows=16384/mem=100KB-16                          104ms ± 2%      106ms ± 8%      ~     (p=0.421 n=5+5)
JoinReader/reqOrdering=true/matchratio=onetoone/lookuprows=16384/mem=unlimited-16                      108ms ± 8%      103ms ± 0%    -4.59%  (p=0.016 n=5+4)
JoinReader/reqOrdering=true/matchratio=onetoone/lookuprows=32768/mem=100KB-16                          211ms ± 5%      209ms ± 4%      ~     (p=0.548 n=5+5)
JoinReader/reqOrdering=true/matchratio=onetoone/lookuprows=32768/mem=unlimited-16                      209ms ± 1%      213ms ± 2%      ~     (p=0.151 n=5+5)
JoinReader/reqOrdering=true/matchratio=onetoone/lookuprows=65536/mem=100KB-16                          420ms ± 1%      417ms ± 2%      ~     (p=0.310 n=5+5)
JoinReader/reqOrdering=true/matchratio=onetoone/lookuprows=65536/mem=unlimited-16                      426ms ± 3%      421ms ± 3%      ~     (p=0.690 n=5+5)
JoinReader/reqOrdering=true/matchratio=onetoone/lookuprows=524288/mem=100KB-16                         3.42s ± 2%      3.39s ± 1%      ~     (p=0.421 n=5+5)
JoinReader/reqOrdering=true/matchratio=onetoone/lookuprows=524288/mem=unlimited-16                     3.40s ± 2%      3.40s ± 3%      ~     (p=0.841 n=5+5)
JoinReader/reqOrdering=true/matchratio=onetofour/lookuprows=1/mem=100KB/parallel=true-16              47.3µs ± 1%     44.6µs ± 3%    -5.76%  (p=0.008 n=5+5)
JoinReader/reqOrdering=true/matchratio=onetofour/lookuprows=1/mem=100KB-16                            47.0µs ± 1%     44.4µs ± 3%    -5.59%  (p=0.008 n=5+5)
JoinReader/reqOrdering=true/matchratio=onetofour/lookuprows=1/mem=unlimited/parallel=true-16          48.0µs ± 2%     44.9µs ± 4%    -6.38%  (p=0.008 n=5+5)
JoinReader/reqOrdering=true/matchratio=onetofour/lookuprows=1/mem=unlimited-16                        47.0µs ± 1%     44.7µs ± 2%    -5.03%  (p=0.008 n=5+5)
JoinReader/reqOrdering=true/matchratio=onetofour/lookuprows=16/mem=100KB-16                            187µs ± 1%      164µs ± 3%   -11.99%  (p=0.008 n=5+5)
JoinReader/reqOrdering=true/matchratio=onetofour/lookuprows=16/mem=unlimited-16                        186µs ± 1%      166µs ± 2%   -10.49%  (p=0.008 n=5+5)
JoinReader/reqOrdering=true/matchratio=onetofour/lookuprows=256/mem=100KB-16                          2.43ms ± 3%     2.06ms ± 2%   -15.33%  (p=0.008 n=5+5)
JoinReader/reqOrdering=true/matchratio=onetofour/lookuprows=256/mem=unlimited-16                      2.46ms ± 1%     2.06ms ± 2%   -16.31%  (p=0.008 n=5+5)
JoinReader/reqOrdering=true/matchratio=onetofour/lookuprows=1024/mem=100KB-16                         9.57ms ± 4%     8.08ms ± 2%   -15.56%  (p=0.008 n=5+5)
JoinReader/reqOrdering=true/matchratio=onetofour/lookuprows=1024/mem=unlimited-16                     9.42ms ± 2%     8.06ms ± 5%   -14.48%  (p=0.008 n=5+5)
JoinReader/reqOrdering=true/matchratio=onetofour/lookuprows=4096/mem=100KB-16                         37.7ms ± 2%     32.3ms ± 3%   -14.25%  (p=0.008 n=5+5)
JoinReader/reqOrdering=true/matchratio=onetofour/lookuprows=4096/mem=unlimited-16                     37.3ms ± 2%     32.2ms ± 1%   -13.60%  (p=0.008 n=5+5)
JoinReader/reqOrdering=true/matchratio=onetofour/lookuprows=8192/mem=100KB-16                         74.2ms ± 1%     64.4ms ± 3%   -13.17%  (p=0.008 n=5+5)
JoinReader/reqOrdering=true/matchratio=onetofour/lookuprows=8192/mem=unlimited-16                     74.0ms ± 2%     64.9ms ± 2%   -12.20%  (p=0.008 n=5+5)
JoinReader/reqOrdering=true/matchratio=onetofour/lookuprows=16384/mem=100KB-16                         151ms ± 2%      127ms ± 3%   -15.50%  (p=0.008 n=5+5)
JoinReader/reqOrdering=true/matchratio=onetofour/lookuprows=16384/mem=unlimited-16                     148ms ± 1%      131ms ± 3%   -11.68%  (p=0.008 n=5+5)
JoinReader/reqOrdering=true/matchratio=onetofour/lookuprows=32768/mem=100KB-16                         297ms ± 2%      257ms ± 5%   -13.53%  (p=0.008 n=5+5)
JoinReader/reqOrdering=true/matchratio=onetofour/lookuprows=32768/mem=unlimited-16                     300ms ± 2%      257ms ± 2%   -14.34%  (p=0.008 n=5+5)
JoinReader/reqOrdering=true/matchratio=onetofour/lookuprows=65536/mem=100KB-16                         589ms ± 2%      514ms ± 3%   -12.67%  (p=0.008 n=5+5)
JoinReader/reqOrdering=true/matchratio=onetofour/lookuprows=65536/mem=unlimited-16                     589ms ± 1%      513ms ± 2%   -12.83%  (p=0.008 n=5+5)
JoinReader/reqOrdering=true/matchratio=onetosixteen/lookuprows=1/mem=100KB/parallel=true-16           56.1µs ± 2%     51.4µs ± 3%    -8.54%  (p=0.008 n=5+5)
JoinReader/reqOrdering=true/matchratio=onetosixteen/lookuprows=1/mem=100KB-16                         55.7µs ± 1%     52.1µs ± 2%    -6.47%  (p=0.008 n=5+5)
JoinReader/reqOrdering=true/matchratio=onetosixteen/lookuprows=1/mem=unlimited/parallel=true-16       55.9µs ± 2%     51.2µs ± 2%    -8.39%  (p=0.008 n=5+5)
JoinReader/reqOrdering=true/matchratio=onetosixteen/lookuprows=1/mem=unlimited-16                     55.5µs ± 0%     52.2µs ± 2%    -5.78%  (p=0.008 n=5+5)
JoinReader/reqOrdering=true/matchratio=onetosixteen/lookuprows=16/mem=100KB-16                         329µs ± 1%      289µs ± 2%   -12.15%  (p=0.008 n=5+5)
JoinReader/reqOrdering=true/matchratio=onetosixteen/lookuprows=16/mem=unlimited-16                     323µs ± 0%      287µs ± 4%   -11.22%  (p=0.008 n=5+5)
JoinReader/reqOrdering=true/matchratio=onetosixteen/lookuprows=256/mem=100KB-16                       8.60ms ± 1%     7.91ms ± 1%    -8.02%  (p=0.008 n=5+5)
JoinReader/reqOrdering=true/matchratio=onetosixteen/lookuprows=256/mem=unlimited-16                   4.52ms ± 2%     3.83ms ± 2%   -15.34%  (p=0.008 n=5+5)
JoinReader/reqOrdering=true/matchratio=onetosixteen/lookuprows=1024/mem=100KB-16                      34.4ms ± 2%     31.5ms ± 2%    -8.35%  (p=0.008 n=5+5)
JoinReader/reqOrdering=true/matchratio=onetosixteen/lookuprows=1024/mem=unlimited-16                  17.7ms ± 1%     15.2ms ± 3%   -13.95%  (p=0.008 n=5+5)
JoinReader/reqOrdering=true/matchratio=onetosixteen/lookuprows=4096/mem=100KB-16                       135ms ± 1%      126ms ± 1%    -6.27%  (p=0.008 n=5+5)
JoinReader/reqOrdering=true/matchratio=onetosixteen/lookuprows=4096/mem=unlimited-16                  70.7ms ± 1%     61.3ms ± 3%   -13.28%  (p=0.008 n=5+5)
JoinReader/reqOrdering=true/matchratio=onetosixteen/lookuprows=8192/mem=100KB-16                       275ms ± 3%      252ms ± 1%    -8.27%  (p=0.008 n=5+5)
JoinReader/reqOrdering=true/matchratio=onetosixteen/lookuprows=8192/mem=unlimited-16                   140ms ± 2%      119ms ± 2%   -14.90%  (p=0.008 n=5+5)
JoinReader/reqOrdering=true/matchratio=onetosixteen/lookuprows=16384/mem=100KB-16                      542ms ± 2%      503ms ± 0%    -7.18%  (p=0.008 n=5+5)
JoinReader/reqOrdering=true/matchratio=onetosixteen/lookuprows=16384/mem=unlimited-16                  279ms ± 2%      239ms ± 2%   -14.44%  (p=0.008 n=5+5)
JoinReader/reqOrdering=true/matchratio=onetosixteen/lookuprows=32768/mem=100KB-16                      1.08s ± 1%      1.02s ± 2%    -4.90%  (p=0.016 n=4+5)
JoinReader/reqOrdering=true/matchratio=onetosixteen/lookuprows=32768/mem=unlimited-16                  566ms ± 3%      482ms ± 2%   -14.87%  (p=0.008 n=5+5)
JoinReader/reqOrdering=true/matchratio=onetothirtytwo/lookuprows=1/mem=100KB/parallel=true-16         67.3µs ± 2%     60.5µs ± 2%   -10.14%  (p=0.008 n=5+5)
JoinReader/reqOrdering=true/matchratio=onetothirtytwo/lookuprows=1/mem=100KB-16                       70.8µs ±14%     60.0µs ± 0%   -15.29%  (p=0.008 n=5+5)
JoinReader/reqOrdering=true/matchratio=onetothirtytwo/lookuprows=1/mem=unlimited/parallel=true-16     73.9µs ±16%     59.8µs ± 1%   -19.11%  (p=0.008 n=5+5)
JoinReader/reqOrdering=true/matchratio=onetothirtytwo/lookuprows=1/mem=unlimited-16                   69.4µs ± 7%     60.0µs ± 3%   -13.62%  (p=0.008 n=5+5)
JoinReader/reqOrdering=true/matchratio=onetothirtytwo/lookuprows=16/mem=100KB-16                       550µs ± 9%      431µs ± 1%   -21.69%  (p=0.008 n=5+5)
JoinReader/reqOrdering=true/matchratio=onetothirtytwo/lookuprows=16/mem=unlimited-16                   537µs ± 7%      431µs ± 3%   -19.79%  (p=0.008 n=5+5)
JoinReader/reqOrdering=true/matchratio=onetothirtytwo/lookuprows=256/mem=100KB-16                     16.6ms ± 5%     14.4ms ±12%   -13.14%  (p=0.016 n=5+5)
JoinReader/reqOrdering=true/matchratio=onetothirtytwo/lookuprows=256/mem=unlimited-16                 7.76ms ± 5%     6.18ms ± 2%   -20.39%  (p=0.008 n=5+5)
JoinReader/reqOrdering=true/matchratio=onetothirtytwo/lookuprows=1024/mem=100KB-16                    63.0ms ± 7%     56.4ms ± 3%   -10.61%  (p=0.008 n=5+5)
JoinReader/reqOrdering=true/matchratio=onetothirtytwo/lookuprows=1024/mem=unlimited-16                31.6ms ± 4%     24.1ms ± 2%   -23.47%  (p=0.008 n=5+5)
JoinReader/reqOrdering=true/matchratio=onetothirtytwo/lookuprows=4096/mem=100KB-16                     265ms ± 7%      220ms ± 1%   -17.11%  (p=0.008 n=5+5)
JoinReader/reqOrdering=true/matchratio=onetothirtytwo/lookuprows=4096/mem=unlimited-16                 121ms ± 8%       96ms ± 3%   -20.56%  (p=0.008 n=5+5)
JoinReader/reqOrdering=true/matchratio=onetothirtytwo/lookuprows=8192/mem=100KB-16                     507ms ± 7%      442ms ± 2%   -12.68%  (p=0.016 n=5+4)
JoinReader/reqOrdering=true/matchratio=onetothirtytwo/lookuprows=8192/mem=unlimited-16                 233ms ±11%      204ms ± 6%   -12.33%  (p=0.008 n=5+5)
JoinReader/reqOrdering=true/matchratio=onetothirtytwo/lookuprows=16384/mem=100KB-16                    1.00s ± 7%      0.91s ± 2%    -9.16%  (p=0.008 n=5+5)
JoinReader/reqOrdering=true/matchratio=onetothirtytwo/lookuprows=16384/mem=unlimited-16                505ms ± 9%      388ms ± 2%   -23.14%  (p=0.008 n=5+5)
JoinReader/reqOrdering=true/matchratio=onetosixtyfour/lookuprows=1/mem=100KB/parallel=true-16         92.4µs ± 6%     79.2µs ± 3%   -14.29%  (p=0.008 n=5+5)
JoinReader/reqOrdering=true/matchratio=onetosixtyfour/lookuprows=1/mem=100KB-16                       89.0µs ±11%     77.9µs ± 2%   -12.45%  (p=0.008 n=5+5)
JoinReader/reqOrdering=true/matchratio=onetosixtyfour/lookuprows=1/mem=unlimited/parallel=true-16     86.4µs ± 3%     78.6µs ± 3%    -9.02%  (p=0.008 n=5+5)
JoinReader/reqOrdering=true/matchratio=onetosixtyfour/lookuprows=1/mem=unlimited-16                   86.4µs ± 2%     78.7µs ± 1%    -8.89%  (p=0.008 n=5+5)
JoinReader/reqOrdering=true/matchratio=onetosixtyfour/lookuprows=16/mem=100KB-16                      1.95ms ± 2%     1.84ms ± 1%    -5.85%  (p=0.008 n=5+5)
JoinReader/reqOrdering=true/matchratio=onetosixtyfour/lookuprows=16/mem=unlimited-16                   836µs ± 1%      734µs ± 1%   -12.21%  (p=0.008 n=5+5)
JoinReader/reqOrdering=true/matchratio=onetosixtyfour/lookuprows=256/mem=100KB-16                     29.0ms ±21%     26.3ms ± 3%    -9.37%  (p=0.032 n=5+5)
JoinReader/reqOrdering=true/matchratio=onetosixtyfour/lookuprows=256/mem=unlimited-16                 12.5ms ± 4%     10.9ms ± 1%   -12.80%  (p=0.008 n=5+5)
JoinReader/reqOrdering=true/matchratio=onetosixtyfour/lookuprows=1024/mem=100KB-16                     112ms ± 8%      106ms ± 1%    -5.94%  (p=0.008 n=5+5)
JoinReader/reqOrdering=true/matchratio=onetosixtyfour/lookuprows=1024/mem=unlimited-16                49.2ms ± 2%     44.0ms ± 2%   -10.48%  (p=0.008 n=5+5)
JoinReader/reqOrdering=true/matchratio=onetosixtyfour/lookuprows=4096/mem=100KB-16                     452ms ± 3%      422ms ± 1%    -6.70%  (p=0.008 n=5+5)
JoinReader/reqOrdering=true/matchratio=onetosixtyfour/lookuprows=4096/mem=unlimited-16                 198ms ± 2%      172ms ± 2%   -12.93%  (p=0.008 n=5+5)
JoinReader/reqOrdering=true/matchratio=onetosixtyfour/lookuprows=8192/mem=100KB-16                     895ms ± 6%      834ms ± 1%    -6.79%  (p=0.008 n=5+5)
JoinReader/reqOrdering=true/matchratio=onetosixtyfour/lookuprows=8192/mem=unlimited-16                 386ms ± 1%      333ms ± 2%   -13.64%  (p=0.008 n=5+5)
JoinReader/reqOrdering=false/matchratio=onetoone/lookuprows=1/mem=unlimited/parallel=true-16          37.6µs ± 3%     38.1µs ± 4%      ~     (p=0.421 n=5+5)
JoinReader/reqOrdering=false/matchratio=onetoone/lookuprows=1/mem=unlimited-16                        37.5µs ± 3%     38.0µs ± 2%      ~     (p=0.151 n=5+5)
JoinReader/reqOrdering=false/matchratio=onetoone/lookuprows=16/mem=unlimited-16                        128µs ± 2%      133µs ± 3%    +4.18%  (p=0.032 n=5+5)
JoinReader/reqOrdering=false/matchratio=onetoone/lookuprows=256/mem=unlimited-16                      1.57ms ± 4%     1.60ms ± 2%      ~     (p=0.222 n=5+5)
JoinReader/reqOrdering=false/matchratio=onetoone/lookuprows=1024/mem=unlimited-16                     6.16ms ± 3%     6.35ms ± 3%    +3.04%  (p=0.032 n=5+5)
JoinReader/reqOrdering=false/matchratio=onetoone/lookuprows=4096/mem=unlimited-16                     25.0ms ± 2%     26.2ms ± 2%    +5.12%  (p=0.008 n=5+5)
JoinReader/reqOrdering=false/matchratio=onetoone/lookuprows=8192/mem=unlimited-16                     50.8ms ± 2%     53.0ms ± 1%    +4.23%  (p=0.008 n=5+5)
JoinReader/reqOrdering=false/matchratio=onetoone/lookuprows=16384/mem=unlimited-16                     116ms ± 2%      119ms ± 2%      ~     (p=0.222 n=5+5)
JoinReader/reqOrdering=false/matchratio=onetoone/lookuprows=32768/mem=unlimited-16                     249ms ± 5%      252ms ± 2%      ~     (p=0.841 n=5+5)
JoinReader/reqOrdering=false/matchratio=onetoone/lookuprows=65536/mem=unlimited-16                     486ms ± 1%      521ms ± 8%    +7.17%  (p=0.008 n=5+5)
JoinReader/reqOrdering=false/matchratio=onetoone/lookuprows=524288/mem=unlimited-16                    4.07s ± 2%      4.19s ± 3%    +2.97%  (p=0.008 n=5+5)
JoinReader/reqOrdering=false/matchratio=onetofour/lookuprows=1/mem=unlimited/parallel=true-16         39.9µs ± 3%     40.6µs ± 3%      ~     (p=0.222 n=5+5)
JoinReader/reqOrdering=false/matchratio=onetofour/lookuprows=1/mem=unlimited-16                       39.6µs ± 2%     39.5µs ± 1%      ~     (p=1.000 n=5+5)
JoinReader/reqOrdering=false/matchratio=onetofour/lookuprows=16/mem=unlimited-16                       158µs ± 3%      146µs ± 1%    -7.85%  (p=0.008 n=5+5)
JoinReader/reqOrdering=false/matchratio=onetofour/lookuprows=256/mem=unlimited-16                     2.02ms ± 2%     1.85ms ± 2%    -8.46%  (p=0.008 n=5+5)
JoinReader/reqOrdering=false/matchratio=onetofour/lookuprows=1024/mem=unlimited-16                    8.01ms ± 3%     7.26ms ± 2%    -9.34%  (p=0.008 n=5+5)
JoinReader/reqOrdering=false/matchratio=onetofour/lookuprows=4096/mem=unlimited-16                    35.2ms ± 3%     32.9ms ± 3%    -6.53%  (p=0.008 n=5+5)
JoinReader/reqOrdering=false/matchratio=onetofour/lookuprows=8192/mem=unlimited-16                    83.3ms ± 3%     76.6ms ± 2%    -8.09%  (p=0.008 n=5+5)
JoinReader/reqOrdering=false/matchratio=onetofour/lookuprows=16384/mem=unlimited-16                    211ms ± 2%      200ms ± 1%    -5.02%  (p=0.008 n=5+5)
JoinReader/reqOrdering=false/matchratio=onetofour/lookuprows=32768/mem=unlimited-16                    487ms ± 4%      472ms ± 3%    -3.19%  (p=0.032 n=5+5)
JoinReader/reqOrdering=false/matchratio=onetofour/lookuprows=65536/mem=unlimited-16                    1.01s ± 3%      0.94s ± 3%    -6.10%  (p=0.008 n=5+5)
JoinReader/reqOrdering=false/matchratio=onetosixteen/lookuprows=1/mem=unlimited/parallel=true-16      45.1µs ± 3%     44.8µs ± 2%      ~     (p=0.841 n=5+5)
JoinReader/reqOrdering=false/matchratio=onetosixteen/lookuprows=1/mem=unlimited-16                    44.1µs ± 2%     44.4µs ± 1%      ~     (p=0.151 n=5+5)
JoinReader/reqOrdering=false/matchratio=onetosixteen/lookuprows=16/mem=unlimited-16                    225µs ± 2%      222µs ± 2%      ~     (p=0.310 n=5+5)
JoinReader/reqOrdering=false/matchratio=onetosixteen/lookuprows=256/mem=unlimited-16                  3.10ms ± 2%     3.03ms ± 2%    -2.12%  (p=0.032 n=5+5)
JoinReader/reqOrdering=false/matchratio=onetosixteen/lookuprows=1024/mem=unlimited-16                 13.0ms ± 1%     12.7ms ± 2%    -1.95%  (p=0.032 n=5+5)
JoinReader/reqOrdering=false/matchratio=onetosixteen/lookuprows=4096/mem=unlimited-16                 68.4ms ± 3%     68.1ms ± 3%      ~     (p=0.841 n=5+5)
JoinReader/reqOrdering=false/matchratio=onetosixteen/lookuprows=8192/mem=unlimited-16                  185ms ± 5%      184ms ± 2%      ~     (p=0.690 n=5+5)
JoinReader/reqOrdering=false/matchratio=onetosixteen/lookuprows=16384/mem=unlimited-16                 561ms ± 5%      546ms ± 3%      ~     (p=0.222 n=5+5)
JoinReader/reqOrdering=false/matchratio=onetosixteen/lookuprows=32768/mem=unlimited-16                 1.41s ± 2%      1.39s ± 3%      ~     (p=0.222 n=5+5)
JoinReader/reqOrdering=false/matchratio=onetothirtytwo/lookuprows=1/mem=unlimited/parallel=true-16    50.8µs ± 1%     51.5µs ± 2%      ~     (p=0.151 n=5+5)
JoinReader/reqOrdering=false/matchratio=onetothirtytwo/lookuprows=1/mem=unlimited-16                  50.0µs ± 2%     50.1µs ± 2%      ~     (p=0.841 n=5+5)
JoinReader/reqOrdering=false/matchratio=onetothirtytwo/lookuprows=16/mem=unlimited-16                  325µs ± 4%      317µs ± 2%      ~     (p=0.151 n=5+5)
JoinReader/reqOrdering=false/matchratio=onetothirtytwo/lookuprows=256/mem=unlimited-16                4.66ms ± 1%     4.58ms ± 2%      ~     (p=0.056 n=5+5)
JoinReader/reqOrdering=false/matchratio=onetothirtytwo/lookuprows=1024/mem=unlimited-16               20.4ms ± 1%     20.3ms ± 1%      ~     (p=0.222 n=5+5)
…